Update external_db_database_transfer.md
Update example DB2 YAML database names for consistency with the DB2 SQL script.
The script uses WPSDB as a common database name throughout the script, and that does not seem to be configurable. The example used to refer to completely different database names. Changed all DB name references to WPSDB.
This change only updates the DB2 section.